// src/webparts/siteCatalog/components/user/FieldUserRenderer.module.scss.d.ts
declare const styles: {
    ellipsis: string;
    user: string;
    relative: string;
    main: string;
    fabricIcon: string;
    actions: string;
    chat: string;
    sections: string;
    section: string;
    header: string;
    chevron: string;
    contactItem: string;
    content: string;
};
export default styles;